// updated homeslider
git-svn-id: http://dev.prestashop.com/svn/v1/branches/1.5.x@11634 b9a71923-0436-4b27-9f14-aed3839534dd

+++ modules/homeslider/HomeSlide.php
@@ -109,8 +109,9 @@ class HomeSlide extends ObjectModel
 		$images = $this->image;
 		foreach ($images as $image)
 		{
-			if ($image && file_exists(dirname(__FILE__).'/images/'.$image))
-				$res &= @unlink(dirname(__FILE__).'/images/'.$image);
+			if (preg_match('/sample/', $image) === 0)
+				if ($image && file_exists(dirname(__FILE__).'/images/'.$image))
+					$res &= @unlink(dirname(__FILE__).'/images/'.$image);
 		}
 
 		$res &= $this->reOrderPositions();
